feat(sharding): Phase A — full ULID conversion (WIP, schema MariaDB-validated)
Convert the whole schema + models to ULID (char(26)) primary/foreign keys.

Migrations (validated on a throwaway MariaDB 10.11 container: 85 tables,
100 FKs, all id/FK cols char(26), morphs->ulidMorphs, framework jobs table
kept bigint):
- bulk id()->ulid('id')->primary(), foreignId->foreignUlid, morphs->ulidMorphs
- crawl re-key website_id ->change() to ulid; mail_transports.oauth_account_id ulid
- personal_access_tokens (Sanctum vendor model) + website_user pivot keep bigint
  surrogate PK (not FK-referenced, not shard-moved); their FKs/morphs are ULID

Models: HasUlids on all 49 (except Setting's string PK; DbNode already had it).

App-code ULID ripple (partial — suite WIP):
- ~430 int/?int id type-hints -> string on params/props/returns
- (int) casts on session('current_website_id'/'impersonator_id'), Auth::id()
- ulid_rows() helper stamps ULIDs on raw DB::table() bulk insert/upsert payloads
- ClientActivityLogger actor/owner resolution returns string ids

Suite: 157 failing (from 40 baseline) — remaining tail is FK-in-test-setup +
a few more id return types + factory/view specifics. Branch is isolated; no
prod/staging DB touched (sqlite tests + ephemeral docker MariaDB only).

+++ app/Http/Controllers/GoogleCapController.php
@@ -67,7 +67,7 @@ class GoogleCapController extends Controller
                 ]);
             }
 
-            if (Auth::check() && in_array((int) Auth::id(), $affectedUserIds->all(), true)) {
+            if (Auth::check() && in_array(Auth::id(), $affectedUserIds->all(), true)) {
                 Auth::logout();
             }
         }
